requirements.txt
click>=8.1.7
pyyaml>=6.0.1
jieba>=0.42.1
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0
pandas>=2.2.0
tqdm>=4.66.0
PyMySQL>=1.1.0
sentence-transformers>=2.5.0
gensim>=4.3.0
networkx>=3.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyvis>=0.3.2
